# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Adega Oak — Liquor Store Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /movimentacoes:
    get:
      operationId: listMovimentacoes
      tags: [movimentacoes]
      summary: Inventory & sales history
      parameters:
        - in: query
          name: tipo
          schema: { type: string, enum: [Entrada, "Saída", All] }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: productid
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, default: 200 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Movimentacao" }
    post:
      operationId: createMovimentacao
      tags: [movimentacoes]
      summary: Register entry (purchase) or exit (sale)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MovimentacaoInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Movimentacao" }
  /movimentacoes/{id}:
    delete:
      operationId: deleteMovimentacao
      tags: [movimentacoes]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
  /vendas/combo:
    post:
      operationId: sellCombo
      tags: [vendas]
      summary: Register a combo sale (debits stock for components when applicable)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ComboVendaInput" }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComboVenda" }
  /vendas/combo/list:
    get:
      operationId: listComboVendas
      tags: [vendas]
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ComboVenda" }
...
  /dashboard/sales-trend:
    get:
      operationId: getSalesTrend
      tags: [dashboard]
      summary: Daily sales for last N days
      parameters:
        - in: query
          name: days
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SalesTrendPoint" }
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Best selling products in given window
      parameters:
        - in: query
          name: days
          schema: { type: integer, default: 30 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TopProduct" }
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 15 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ActivityItem" }
  /funcionarios:
    get:
      operationId: listFuncionarios
      tags: [funcionarios]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Funcionario" }
    post:
      operationId: createFuncionario
      tags: [funcionarios]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FuncionarioInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Funcionario" }
  /funcionarios/{id}:
    delete:
      operationId: deleteFuncionario
      tags: [funcionarios]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Movimentacao:
      type: object
      required: [id, data, tipo, productid, produto, quantidade, responsavel, valor_unitario]
      properties:
        id: { type: integer }
        data: { type: string, format: date-time }
        tipo: { type: string, enum: [Entrada, "Saída"] }
        tipo_venda: { type: string, enum: [Varejo, Atacado], nullable: true }
        productid: { type: integer }
        produto: { type: string }
        quantidade: { type: integer }
        responsavel: { type: string }
        saida:
          type: string
          nullable: true
        valor_unitario: { type: number }
        valor_total: { type: number }
        observacoes:
          type: string
          nullable: true
    MovimentacaoInput:
      type: object
      required: [tipo, productid, quantidade, responsavel, valor_unitario]
      properties:
        tipo: { type: string, enum: [Entrada, "Saída"] }
        tipo_venda: { type: string, enum: [Varejo, Atacado], default: Varejo }
        productid: { type: integer }
        quantidade: { type: integer, minimum: 1 }
        responsavel: { type: string, minLength: 1 }
        saida:
          type: string
          nullable: true
        valor_unitario: { type: number, minimum: 0 }
        observacoes:
          type: string
          nullable: true
...
    ComboVenda:
      type: object
      required: [venda_id, combo_id, quantidade, preco_unitario, preco_total, data_venda]
      properties:
        venda_id: { type: integer }
        combo_id: { type: integer }
        nome:
          type: string
          nullable: true
        quantidade: { type: integer }
        preco_unitario: { type: number }
        preco_total: { type: number }
        data_venda: { type: string, format: date-time }
        responsavel:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        tipo_movimento: { type: string }
    ComboVendaInput:
      type: object
      required: [combo_id, quantidade, responsavel]
      properties:
        combo_id: { type: integer }
        quantidade: { type: integer, minimum: 1 }
        preco_unitario:
          type: number
          minimum: 0
          nullable: true
          description: "Optional override; otherwise uses combo's preco_venda"
        responsavel: { type: string }
        observacoes:
          type: string
          nullable: true
...
    SalesTrendPoint:
      type: object
      required: [date, vendas, lucro, transacoes]
      properties:
        date: { type: string, format: date }
        vendas: { type: number }
        lucro: { type: number }
        transacoes: { type: integer }
    TopProduct:
      type: object
      required: [productid, produto, quantidade_vendida, receita]
      properties:
        productid: { type: integer }
        produto: { type: string }
        quantidade_vendida: { type: integer }
        receita: { type: number }
    ActivityItem:
      type: object
      required: [id, kind, title, subtitle, amount, when]
      properties:
        id: { type: string }
        kind: { type: string, enum: [venda, entrada, despesa, combo] }
        title: { type: string }
        subtitle: { type: string }
        amount: { type: number }
        when: { type: string, format: date-time }
    Funcionario:
      type: object
      required: [id, username]
      properties:
        id: { type: integer }
        username: { type: string }
        ativo: { type: boolean }
    FuncionarioInput:
      type: object
      required: [username]
      properties:
        username: { type: string, minLength: 2, maxLength: 50 }
...