# lib/api-zod/src/generated/types/comboVenda.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adega Oak — Liquor Store Management API
 * OpenAPI spec version: 0.1.0
 */

export interface ComboVenda {
  venda_id: number;
  combo_id: number;
  nome?: string | null;
  quantidade: number;
  preco_unitario: number;
  preco_total: number;
  data_venda: Date;
  responsavel?: string | null;
  observacoes?: string | null;
  tipo_movimento?: string;
}