# lib/api-zod/src/generated/types/funcionarioInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adega Oak — Liquor Store Management API
 * OpenAPI spec version: 0.1.0
 */

export interface FuncionarioInput {
  /**
   * @minLength 2
   * @maxLength 50
   */
  username: string;
}