# lib/api-zod/src/generated/types/movimentacaoInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adega Oak — Liquor Store Management API
 * OpenAPI spec version: 0.1.0
 */
import type { MovimentacaoInputTipo } from "./movimentacaoInputTipo";
import type { MovimentacaoInputTipoVenda } from "./movimentacaoInputTipoVenda";

export interface MovimentacaoInput {
  tipo: MovimentacaoInputTipo;
  tipo_venda?: MovimentacaoInputTipoVenda;
  productid: number;
  /** @minimum 1 */
  quantidade: number;
  /** @minLength 1 */
  responsavel: string;
  saida?: string | null;
  /** @minimum 0 */
  valor_unitario: number;
  observacoes?: string | null;
}