# lib/api-zod/src/generated/types/salesTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adega Oak — Liquor Store Management API
 * OpenAPI spec version: 0.1.0
 */

export interface SalesTrendPoint {
  date: Date;
  vendas: number;
  lucro: number;
  transacoes: number;
}